index: refactor selector handling
Unify empty string and nil by nilifying the selector when the empty string is
received. This allow to extract selector handling in two separate functions, one
for the empty selector and one for the non-empty one, that can be reused in
mount and handle_event.
Also, remove :where assign since it was useless.